Separate UA parsing into its own file
